requirements.txt
# Console output
rich

# Config file
pyyaml
jsonschema
environs

# Server
fastapi[all]
fastapi-socketio
uvicorn[standard]

# Database
# > Do NOT install `bson` here as `pymongo` installs its own `bson`.
pymongo
pydantic

# Security
python-jose[cryptography]
passlib[bcrypt]
